feat(map-popover): refine POI modal display
- hide empty rows
- preserve image ratio on bigger images
- use full height on detailed view
- more spacing between rows on detailed view
